Split Entity.php
