+++ process/Model/Wikidata/LanguageValue.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Model\Wikidata;
+
+class LanguageValue
+{
+    public string $language;
+    public string $value;
+}
